repository de proveedor y reparacion.

+++ src/main/java/org/springframework/samples/petclinic/model/Reparacion.java
@@ -11,7 +11,7 @@ public class Reparacion extends BaseEntity{
 	@Column(name = "duracion")
 	private Integer duracion;
 	@Column(name = "precio")
-	private Integer precio;
+	private Float precio;
 	@Column(name = "tipoReparacion")
 	private TipoReparacion tipoReparacion;
 	
@@ -21,10 +21,10 @@ public class Reparacion extends BaseEntity{
 	public void setDuracion(Integer duracion) {
 		this.duracion = duracion;
 	}
-	public Integer getPrecio() {
+	public Float getPrecio() {
 		return precio;
 	}
-	public void setPrecio(Integer precio) {
+	public void setPrecio(Float precio) {
 		this.precio = precio;
 	}
 	public TipoReparacion getTipoReparacion() {
